% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finite_non_rand2.r
\name{finite_non_rand2}
\alias{finite_non_rand2}
\title{Simulating single locus, two allele Mendelian populations with 
finite population sizes and non-random mating over many generations}
\usage{
finite_non_rand2(
  outcrossingrate_nr,
  p1g2_nr,
  p2g2_nr,
  alleles,
  ND_nr,
  generations_nr,
  n_rep
)
}
\arguments{
\item{outcrossingrate_nr}{Outcrossing rate}

\item{p1g2_nr}{Allele 1 frequency}

\item{p2g2_nr}{Allele 2 frequency}

\item{alleles}{Allele types (use 1 and 2)}

\item{ND_nr}{Population size}

\item{generations_nr}{Number of generations}

\item{n_reps}{Number simulation replicates}
}
\description{
\code{finite_non_rand2} simulates finite Mendelian populations with non-random mating over many generations. 
Specify initial allele frequencies, population size, number of generations, outcrossing rate, and number of replicates.
The simulation produces each generation of the population by randomly selecting 2N (ND) alleles for N (ND) individuals from a gamete pool based on 
the allele frequencies.
}
\details{
\code{finite_non_rand} returns three plots: 
1) Allele 1 frequency versus number of generations, 2) Observed and expected heterozygosity (expected based on
allele frequencies and drift) versus number of generations, and 3) Wright's F with a \link{loess} regression best fit line.
It also returns Fisher's exact tests (\link{fisher.test}) comparing the observed genotype counts to the expected
and initial values.
}
\examples{
finite_non_rand(0.5,0.3,0.7,c(1,2),25,100,100)
}
