% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finite_pop.r
\name{finite_pop}
\alias{finite_pop}
\title{Simulating single locus, two allele Mendelian populations
with finite population sizes over many generations}
\usage{
finite_pop(p1g2, p2g2, alleles, ND, generations)
}
\arguments{
\item{p1g2}{Allele 1 frequency}

\item{p2g2}{Allele 2 frequency}

\item{alleles}{allele types (use 1 and 2)}

\item{ND}{Population size}

\item{generations}{number of generations}
}
\description{
\code{finite_pop} simulates Mendelian populations over many generations. Specify initial allele
frequencies, population size, and the number of generations.  The simulation produces each generation of
the population by randomly selecting 2N (ND) alleles for N (ND) individuals from a gamete pool based on 
allele frequencies.
}
\details{
\code{finite_pop} returns three plots: 
1) Allele 1 frequency versus number of generations, 2) Observed and expected heterozygosity versus number
of generations, and 3) Wright's F with a \link{loess} regression best fit line.
It also returns Fisher's exact tests (\link{fisher.test}) comparing the observed genotype counts to the expected
and initial values.
}
\examples{
finite_pop(0.3,0.7,c(1,2),100,100)
}
