% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finite_selection.R
\name{finite_selection}
\alias{finite_selection}
\title{Simulating single locus, two allele Mendelian populations with
finite population sizes and natural selection over many generations}
\usage{
finite_selection(s, h, p1g2_nr, p2g2_nr, alleles, ND_nr, generations_nr, n_rep)
}
\arguments{
\item{s}{Selection coefficient}

\item{h}{Dominance coefficient}

\item{p1g2_nr}{Allele 1 frequency}

\item{p2g2_nr}{Allele 2 frequency}

\item{alleles}{Allele types (use 1 and 2)}

\item{ND_nr}{Population size}

\item{generations_nr}{Number of generations}

\item{n_reps}{Number simulation replicates}
}
\description{
\code{finite_selection} simulates finite Mendelian populations with selection over many generations.
Specify initial allele frequencies, population size, number of generations, selection coefficient, dominance coefficient, and number of replicates.
The simulation produces each generation of the population by randomly selecting 2N (ND) alleles for N (ND) individuals from a gamete pool based on
the allele frequencies as determined by the relative fitnesses of the genotypes.
}
\details{
\code{finite_selection} returns three plots:
1) Allele 1 frequency versus number of generations, 2) Observed and expected heterozygosity (expected based on
allele frequencies and drift) versus number of generations, and 3) Wright's F with a \link{loess} regression best fit line.
It also returns Fisher's exact tests (\link{fisher.test}) comparing the observed genotype counts to the expected
and initial values.
}
\examples{
finite_selection(0.5,0.5,0.3,0.7,c(1,2),25,100,100)
}
