% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lp.r
\name{lp}
\alias{lp}
\title{Simulating (large) Mendelian populations}
\usage{
lp(p1, p2, alleles, N)
}
\arguments{
\item{p1}{Allele 1 frequency}

\item{p2}{Allele 2 frequency}

\item{alleles}{allele types (use 1 and 2)}

\item{N}{Population size}
}
\description{
\code{lp} simulates Mendelian populations. Specify allele
frequencies and population size.  The simulation generates a population with N individuals 
and 2N alleles by randomly selecting 2N alleles for N individuals from a gamete pool based on 
allele frequencies.
}
\details{
\code{lp} returns the allele frequencies, observed and expected heterozygosities,
Wright's F, and Fisher's exact tests (\link{fisher.test}) comparing the observed genotype counts to the expected
and initial values.
}
\examples{
lp(0.3,0.7,c(1,2),100)
}
