% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/non_rand_mating.r
\name{non_rand_mating}
\alias{non_rand_mating}
\title{Simulating non-random mating Mendelian populations}
\usage{
non_rand_mating(outcrossingrate, p1, p2, alleles, N)
}
\arguments{
\item{outcrossingrate}{probability an individual mates with a randomly selected individual from the parent population}

\item{p1}{Allele 1 frequency}

\item{p2}{Allele 2 frequency}

\item{alleles}{allele types (use 1 and 2)}

\item{N}{Population size}
}
\description{
\code{non_rand_mating} simulates non-random mating Mendelian populations by specifying allele
frequencies, population size, and outcrossing rate.  The simulation generates a parent population with N individuals 
with 2N alleles by randomly sampling from a gamete pool based on the specified allele frequencies.  Then parents are randomly selected to mate where the
outcrossing rate determines whether they mate with themselves or another randomly selected individual.
}
\details{
\code{non_rand_mating} returns the allele frequencies, observed and expected heterozygosities,
Wright's F, and Fisher's exact (\link{fisher.test}) tests comparing the observed genotype counts to the expected
and initial values.
}
\examples{
non_rand_mating(0.5,0.3,0.7,c(1,2),1000)
}
